add hoc with favorite handler
