+++ src/components/main-page/hotel-card/hotel-card.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import PropTypes from "prop-types";
 import BookmarkButton from "../../bookmark-button/bookmark-button";
-import Operation from './../../../store/actions/async-actions/async-actions';
-import {connect} from 'react-redux';
 import {Link} from "react-router-dom";
+import withFavoriteHandler from '../../../hocs/with-favorite-handler/with-favorite-handler';
 
+const BookmarkButtonWrapped = withFavoriteHandler(BookmarkButton);
 function Premium() {
   return (
     <div className="place-card__mark">
@@ -14,7 +14,7 @@ function Premium() {
 }
 
 function HotelCard(props) {
-  const {onHover, offer, history, favoriteHotelHandler, view = `cities`} = props;
+  const {onHover, offer, history, view = `cities`} = props;
   const answer = offer;
   const {isPremium, isFavorite, cost, rating, name, type, previewImage, id} = offer;
 
@@ -42,9 +42,10 @@ function HotelCard(props) {
             <b className="place-card__price-value">&euro;{cost}</b>
             <span className="place-card__price-text">&#47;&nbsp;night</span>
           </div>
-          <BookmarkButton
-            isBookmarkAdded={isFavorite}
-            onClick={_isFavoriteHotelHandler}
+          <BookmarkButtonWrapped
+            isFavorite={isFavorite}
+            history = {history}
+            id = {id}
           />
         </div>
         <div className="place-card__rating rating">
@@ -62,13 +63,6 @@ function HotelCard(props) {
       </div>
     </article>
   );
-
-  function _isFavoriteHotelHandler(evt) {
-    evt.preventDefault();
-    let status = isFavorite ? 0 : 1;
-    favoriteHotelHandler(id, status, history);
-  }
-
 }
 
 HotelCard.propTypes = {
@@ -100,13 +94,5 @@ HotelCard.propTypes = {
   view: PropTypes.string
 };
 
-const mapDispatchToProps = (dispatch) => ({
-  favoriteHotelHandler: (id, status, history) => {
-    dispatch(Operation.favoriteHotelHandler(id, status, history));
-  }
-});
 
-
-export {HotelCard};
-
-export default connect(null, mapDispatchToProps)(HotelCard);
+export default HotelCard;
